pyproject.toml
[tool.poetry]
name = "tidydir"
version = "0.1.0"
description = "tidydir organizes files with dates in their names into directories."
authors = ["mozkzki <mozkzki@gmail.com>"]
maintainers = ["mozkzki <mozkzki@gmail.com>"]
readme = "README.md"
license = "MIT"
homepage = "https://github.com/mozkzki/tidydir"
repository = "https://github.com/mozkzki/tidydir"
documentation = "https://github.com/mozkzki/tidydir/wiki"
keywords = ['tool', 'rename', 'organize']
classifiers=[
    'Development Status :: 3 - Alpha',
    'License :: OSI Approved :: MIT License',
    'Topic :: Software Development :: Libraries',
    'Intended Audience :: Developers',
    'Programming Language :: Python :: 3.7',
    'Programming Language :: Python :: 3.8',
    'Programming Language :: Python :: 3.9',
]

[tool.poetry.dependencies]
python = "^3.7"
python-dotenv = "^0.19.2"
requests = "2.26.0"
pillow = "^9.3.0"

[tool.poetry.dev-dependencies]
flake8 = "4.0.1"
mypy = "0.910"
pytest = "6.2.5"
black = "21.11b1"
pytest-cov = "3.0.0"
coverage = "6.1.2"
isort = "5.10.1"
Sphinx = "^4.3.0"
sphinx-rtd-theme = "^1.0.0"
types-requests = "^2.26.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
tidydir = 'tidydir.cli:run'
